+++ packages/url_launcher/url_launcher_macos/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.1.0-nullsafety.1
+
+* Bump SDK to support null safety.
+
+# 0.1.0-nullsafety
+
+* Migrate to null safety.
+
+## 0.0.2+1
+
+* Update Flutter SDK constraint.
+
 ## 0.0.2
 
 * Update integration test examples to use `testWidgets` instead of `test`.